// source/frontend/src/models/api/generated/ApiGen_CodeTypes_LeaseStatusTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_LeaseStatusTypes {
  ACTIVE = 'ACTIVE',
  ARCHIVED = 'ARCHIVED',
  DISCARD = 'DISCARD',
  DRAFT = 'DRAFT',
  DUPLICATE = 'DUPLICATE',
  EXPIRED = 'EXPIRED',
  INACTIVE = 'INACTIVE',
  TERMINATED = 'TERMINATED',
}