// source/frontend/src/models/api/UserOverrideCode.ts
export enum UserOverrideCode {
  ADD_PROPERTY_TO_INVENTORY = 'ADD_PROPERTY_TO_INVENTORY',
  ADD_LOCATION_TO_PROPERTY = 'ADD_LOCATION_TO_PROPERTY',
  UPDATE_REGION = 'UPDATE_REGION',
  PROPERTY_OF_INTEREST_TO_INVENTORY = 'PROPERTY_OF_INTEREST_TO_INVENTORY',
  CONTRACTOR_SELFREMOVED = 'CONTRACTOR_SELFREMOVED',
  PRODUCT_REUSE = 'PRODUCT_REUSE',
  DISPOSING_PROPERTY_NOT_INVENTORIED = 'DISPOSING_PROPERTY_NOT_INVENTORIED',
  MANAGING_PROPERTY_NOT_INVENTORIED = 'MANAGING_PROPERTY_NOT_INVENTORIED',
  DISPOSITION_FILE_FINAL_STATUS = 'DISPOSITION_FILE_FINAL_STATUS',
  DISPOSE_OF_PROPERTIES = 'DISPOSE_OF_PROPERTIES',
  DELETE_COMPLETED_TAKE = 'DELETE_COMPLETED_TAKE',
  DELETE_LAST_TAKE = 'DELETE_LAST_TAKE',
  DELETE_TAKE_ACTIVE_DISPOSITION = 'DELETE_TAKE_ACTIVE_DISPOSITION',
  UPDATE_SUBFILES_PROJECT_PRODUCT = 'UPDATE_SUBFILES_PROJECT_PRODUCT',
  COMMENCEMENT_DATE_OVERLAP_EXPIRY_DATE = 'COMMENCEMENT_DATE_OVERLAP_EXPIRY_DATE',
}